+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-18 17:57:27
-// Commit: refactor(i18n): optimize data flow between domain and data layers
+// Updated at: 2026-05-18 18:18:40
+// Commit: feat(i18n): add bi-directional event bus handler for i18n
 
 class i18nComponentState {
-  final String commitRef = '1264';
-  final String lastMessage = 'refactor(i18n): optimize data flow between domain and data layers';
-  final DateTime timestamp = DateTime.parse('2026-05-18T17:57:27+05:00');
+  final String commitRef = '1267';
+  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
+  final DateTime timestamp = DateTime.parse('2026-05-18T18:18:40+05:00');
 }